pyproject.toml
[tool.poetry]
name = "code-dataset-builder"
version = "1.0.0"
description = ""
authors = ["notoxicpeople"]
readme = "README.md"
packages = [    
    { include = "dataset", from = "src" },
    { include = "util", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
datasets = "^2.18.0"
smart-open = "^7.0.4"
transformers = "^4.39.1"
boto3 = "^1.34.72"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
pyproject-flake8 = "^6.1.0"
black = "^24.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
line-length = 200

[tool.isort]
line_length = 200
profile = "black"

[tool.flake8]
max-line-length = 200
max-complexity = 18
ignore = "ANN101,ANN204,W503,E501,E116,E266,E722,"
exclude = "__pycache__,.mypy_cache,.pytest_cache,.venv,old,build,dist"

[tool.mypy]
python_version = "3.11"
no_strict_optional = true
ignore_missing_imports = true
check_untyped_defs = true


